Move abfs-vector3_order from source_base to module_ri
