+++ source/source_io/module_restart/restart.h
@@ -3,7 +3,7 @@
 
 #include <string>
 #include "source_base/global_function.h"
-#include "source_base/abfs-vector3_order.h"
+#include "source_lcao/module_ri/abfs-vector3_order.h"
 #ifdef __EXX
 #include <RI/global/Tensor.h>
 #endif
