+++ source/source_lcao/module_ri/LRI_CV.h
@@ -9,7 +9,7 @@
 #include "Matrix_Orbs11.h"
 #include "Matrix_Orbs21.h"
 #include "source_basis/module_ao/ORB_atomic_lm.h"
-#include "source_base/abfs-vector3_order.h"
+#include "abfs-vector3_order.h"
 #include "source_base/element_basis_index.h"
 
 #include <RI/global/Tensor.h>
