+++ source/source_lcao/module_ri/serialization_cereal.h
@@ -10,7 +10,7 @@
 #include <cereal/types/functional.hpp>
 
 #include "../../source_base/vector3.h"
-#include "../../source_base/abfs-vector3_order.h"
+#include "abfs-vector3_order.h"
 #include "../../source_base/matrix.h"
