+++ source/source_lcao/module_ri/test/CMakeLists.txt
@@ -11,4 +11,9 @@ AddTest(
   LIBS base ${math_libs} device parameter
   SOURCES ri_cv_io_test.cpp
 )
+AddTest(
+  TARGET MODULE_RI_abfs-vector3_order
+  LIBS parameter
+  SOURCES abfs-vector3_order_test.cpp ../abfs-vector3_order.cpp
+)
 install(DIRECTORY support DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
