+++ source/source_lcao/module_ri/test/abfs-vector3_order_test.cpp
@@ -1,4 +1,4 @@
-#include "source_base/abfs-vector3_order.h"
+#include "source_lcao/module_ri/abfs-vector3_order.h"
 #include "source_base/vector3.h"
 #include "gtest/gtest.h"
 #include "source_lcao/module_ri/abfs.h"
