+++ source/source_lcao/module_rt/td_folding.h
@@ -1,7 +1,7 @@
 #ifndef TD_FOLDING_H
 #define TD_FOLDING_H
 #include "source_lcao/module_hcontainer/hcontainer.h"
-#include "source_base/abfs-vector3_order.h"
+#include "source_lcao/module_ri/abfs-vector3_order.h"
 
 namespace module_rt{
 // folding HR to hk, for hybrid gauge
