+++ source/source_lcao/spar_u.h
@@ -1,7 +1,7 @@
 #ifndef SPARSE_FORMAT_U_H 
 #define SPARSE_FORMAT_U_H
 
-#include "source_base/abfs-vector3_order.h"
+#include "source_lcao/module_ri/abfs-vector3_order.h"
 #include "source_cell/module_neighbor/sltk_atom_arrange.h"
 #include "source_cell/module_neighbor/sltk_grid_driver.h"
 #include "source_lcao/module_dftu/dftu.h" // mohan add 20251107
